examples/objectives-report/requirements.txt
pygal==3.0.4
python-dateutil==2.9.0
